apps/api/requirements.txt
flask==3.0.0
PyJWT==2.8.0
Flask-CORS==4.0.1
